remittance

+++ Applications/App/Modules/Pannel/Views/listeproduit.php
@@ -3,8 +3,7 @@
 
         <div class="white-box">
             <h3 class="box-title">Liste des Produits</h3>
-            <button type="button" class="btn btn-primary" id="button" data-toggle="modal" data-target="#addAgence"
-                data-whatever="@mdo"><i class="fa fa-plus"> Ajouter</i></button> <br /> <br />
+            <button type="button" class="btn btn-primary" id="button" data-toggle="modal" data-target="#addAgence" data-whatever="@mdo"><i class="fa fa-plus"> Ajouter</i></button> <br /> <br />
             <div class="table-responsive">
                 <table id="dataTable" class="display nowrap" cellspacing="0" width="100%">
                     <thead>
@@ -17,55 +16,43 @@
                     </thead>
                     <tbody>
                         <?php foreach ($ListeProduit as $key => $value) { ?>
-                        <tr>
-                            <td><?= $value['RefProduit']; ?></td>
-                            <td><?= $value['NameProduit']; ?></td>
-                            <td><?= $value['NameBanque']; ?></td>
-                            <td>
-                                <a href="/Pannel/Produit/delete/<?= $value['RefProduit']; ?>"
-                                    class="btn btn-xs btn-danger"
-                                    onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet élément ?');"><i
-                                        class="fa fa-trash"></i></a>
-                                <a class="btn btn-xs  btn-warning" data-toggle="modal"
-                                    data-target="#ChmodProduit-<?= $value['RefProduit']; ?>" data-whatever="@mdo"><i
-                                        class="fa fa-check"></i></a>
+                            <tr>
+                                <td><?= $value['RefProduit']; ?></td>
+                                <td><?= $value['NameProduit']; ?></td>
+                                <td><?= $value['NameBanque']; ?></td>
+                                <td>
+                                    <a href="/Pannel/Produit/delete/<?= $value['RefProduit']; ?>" class="btn btn-xs btn-danger" onclick="return confirm('Êtes-vous sûr de vouloir supprimer cet élément ?');"><i class="fa fa-trash"></i></a>
+                                    <a class="btn btn-xs  btn-warning" data-toggle="modal" data-target="#ChmodProduit-<?= $value['RefProduit']; ?>" data-whatever="@mdo"><i class="fa fa-check"></i></a>
 
-                            </td>
-                        </tr>
-                        <div class="modal fade" id="ChmodProduit-<?= $value['RefProduit']; ?>" tabindex="-1"
-                            role="dialog" aria-labelledby="exampleModalLabel1">
-                            <div class="modal-dialog" role="document">
-                                <div class="modal-content">
-                                    <div class="modal-header">
-                                        <button type="button" class="close" data-dismiss="modal"
-                                            aria-label="Close"><span aria-hidden="true">&times;</span></button>
-                                    </div>
-                                    <form role="form" method="post" action="">
-                                        <input type="hidden" class="form-control" name="RefProduit"
-                                            value="<?= $value['RefProduit']; ?>">
-                                        <div class="modal-body">
-                                            <?php foreach ($ListeCaisse as $key => $caisse) { ?>
-                                            <div class="checkbox checkbox-success checkbox-circle">
-                                                <input id="checkbox-10" type="checkbox" name="RefCaisse[]" multiple=""
-                                                    value="<?= $caisse['RefCaisse']; ?>"
-                                                    <?php if ($produits[$caisse['RefCaisse']][$value['RefProduit']] == $caisse['RefCaisse']) { ?>
-                                                    checked="" <?php } ?> />
-                                                <label for="checkbox-10">
-                                                    <?= $caisse['NameCaisse'] . " | " . $caisse['NameAgency']; ?></label>
+                                </td>
+                            </tr>
+                            <div class="modal fade" id="ChmodProduit-<?= $value['RefProduit']; ?>" tabindex="-1" role="dialog" aria-labelledby="exampleModalLabel1">
+                                <div class="modal-dialog" role="document">
+                                    <div class="modal-content">
+                                        <div class="modal-header">
+                                            <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
+                                        </div>
+                                        <form role="form" method="post" action="">
+                                            <input type="hidden" class="form-control" name="RefProduit" value="<?= $value['RefProduit']; ?>">
+                                            <div class="modal-body">
+                                                <?php foreach ($ListeCaisse as $key => $caisse) { ?>
+                                                    <div class="checkbox checkbox-success checkbox-circle">
+                                                        <input id="checkbox-10" type="checkbox" name="RefCaisse[]" multiple="" value="<?= $caisse['RefCaisse']; ?>" <?php if ($produits[$caisse['RefCaisse']][$value['RefProduit']] == $caisse['RefCaisse']) { ?> checked="" <?php } ?> />
+                                                        <label for="checkbox-10">
+                                                            <?= $caisse['NameCaisse'] . " | " . $caisse['NameAgency']; ?></label>
+                                                    </div>
+                                                <?php } ?>
+
                                             </div>
-                                            <?php } ?>
 
-                                        </div>
-
-                                        <div class="modal-footer">
-                                            <button type="button" class="btn btn-default"
-                                                data-dismiss="modal">Fermer</button>
-                                            <button type="submit" class="btn btn-primary">Valider</button>
-                                        </div>
-                                    </form>
+                                            <div class="modal-footer">
+                                                <button type="button" class="btn btn-default" data-dismiss="modal">Fermer</button>
+                                                <button type="submit" class="btn btn-primary">Valider</button>
+                                            </div>
+                                        </form>
+                                    </div>
                                 </div>
                             </div>
-                        </div>
                         <?php } ?>
                     </tbody>
                 </table>
@@ -77,8 +64,7 @@
     <div class="modal-dialog" role="document">
         <div class="modal-content">
             <div class="modal-header">
-                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span
-                        aria-hidden="true">&times;</span></button>
+                <button type="button" class="close" data-dismiss="modal" aria-label="Close"><span aria-hidden="true">&times;</span></button>
             </div>
             <form role="form" method="post" action="">
                 <div class="modal-body">
@@ -91,7 +77,7 @@
                     <select name="RefBanque" class="form-control">
                         <option value="">Veuillez Choisir la Banque</option>
                         <?php foreach ($ListeBanque as $key => $value) { ?>
-                        <option value="<?= $value['RefBanque']; ?>"><?= $value['NameBanque']; ?></option>
+                            <option value="<?= $value['RefBanque']; ?>"><?= $value['NameBanque']; ?></option>
                         <?php   } ?>
                     </select>
                 </div>
